ovs: move event to events mod
