+++ src/events/mod.rs
@@ -23,4 +23,5 @@ pub(crate) mod file;
 
 pub(crate) mod ct;
 pub(crate) mod nft;
+pub(crate) mod ovs;
 pub(crate) mod skb;
