+++ src/module/ovs/mod.rs
@@ -8,9 +8,6 @@ pub(crate) mod ovs;
 // Re-export ovs.rs
 pub(crate) use ovs::*;
 
-pub(crate) mod event;
-pub(crate) use event::*;
-
 mod bpf;
 mod hooks {
     pub(super) mod kernel_enqueue {
